refactored queue
